refactor(app): rename delete to remove for consistency with instance and secret [APIAPEX-2823]
